+++ src/pyscipopt/scip.pyx
@@ -288,6 +288,11 @@ cdef class Event:
         cdef SCIP_NODE* node = SCIPeventGetNode(self.event)
         return Node.create(node)
 
+    def getRow(self):
+        """gets row for a row event"""
+        cdef SCIP_ROW* row = SCIPeventGetRow(self.event)
+        return Row.create(row)
+
 cdef class Column:
     """Base class holding a pointer to corresponding SCIP_COL"""
 
@@ -344,6 +349,11 @@ cdef class Row:
         row = Row()
         row.scip_row = sciprow
         return row
+
+    property name:
+        def __get__(self):
+            cname = bytes( SCIProwGetName(self.scip_row) )
+            return cname.decode('utf-8')
 
     def getLhs(self):
         """returns the left hand side of row"""
@@ -503,8 +513,9 @@ cdef class Node:
         return node
 
     def getParent(self):
-        """Retrieve parent node."""
-        return Node.create(SCIPnodeGetParent(self.scip_node))
+        """Retrieve parent node (or None if the node has no parent node)."""
+        cdef SCIP_NODE* parent = SCIPnodeGetParent(self.scip_node)
+        return Node.create(parent) if parent != NULL else None
 
     def getNumber(self):
         """Retrieve number of node."""
@@ -856,7 +867,7 @@ cdef class Model:
         return SCIPgetPresolvingTime(self._scip)
 
     def getNLPIterations(self):
-        """Retrieve the current number of LP iterations."""
+        """Retrieve the total number of LP iterations so far."""
         return SCIPgetNLPIterations(self._scip)
 
     def getNNodes(self):
@@ -3313,16 +3324,15 @@ cdef class Model:
         cdef int npriolpcands
         cdef int nfracimplvars
 
-        ncands = SCIPgetNLPBranchCands(self._scip)
-        cdef SCIP_VAR** lpcands = <SCIP_VAR**> malloc(ncands * sizeof(SCIP_VAR*))
-        cdef SCIP_Real* lpcandssol = <SCIP_Real*> malloc(ncands * sizeof(SCIP_Real))
-        cdef SCIP_Real* lpcandsfrac = <SCIP_Real*> malloc(ncands * sizeof(SCIP_Real))
+        cdef SCIP_VAR** lpcands
+        cdef SCIP_Real* lpcandssol
+        cdef SCIP_Real* lpcandsfrac
 
         PY_SCIP_CALL(SCIPgetLPBranchCands(self._scip, &lpcands, &lpcandssol, &lpcandsfrac,
                                           &nlpcands, &npriolpcands, &nfracimplvars))
 
-        return ([Variable.create(lpcands[i]) for i in range(ncands)], [lpcandssol[i] for i in range(ncands)],
-                [lpcandsfrac[i] for i in range(ncands)], nlpcands, npriolpcands, nfracimplvars)
+        return ([Variable.create(lpcands[i]) for i in range(nlpcands)], [lpcandssol[i] for i in range(nlpcands)],
+                [lpcandsfrac[i] for i in range(nlpcands)], nlpcands, npriolpcands, nfracimplvars)
 
 
     def branchVar(self, variable):
@@ -3332,9 +3342,9 @@ cdef class Model:
         :return: tuple(downchild, eqchild, upchild) of Nodes of the left, middle and right child.
 
         """
-        cdef SCIP_NODE* downchild = <SCIP_NODE*> malloc(sizeof(SCIP_NODE))
-        cdef SCIP_NODE* eqchild = <SCIP_NODE*> malloc(sizeof(SCIP_NODE))
-        cdef SCIP_NODE* upchild = <SCIP_NODE*> malloc(sizeof(SCIP_NODE))
+        cdef SCIP_NODE* downchild
+        cdef SCIP_NODE* eqchild
+        cdef SCIP_NODE* upchild
 
         PY_SCIP_CALL(SCIPbranchVar(self._scip, (<Variable>variable).scip_var, &downchild, &eqchild, &upchild))
         return Node.create(downchild), Node.create(eqchild), Node.create(upchild)
@@ -3349,12 +3359,12 @@ cdef class Model:
                     if branch variable is integer
 
         """
-        cdef SCIP_NODE* downchild = <SCIP_NODE*> malloc(sizeof(SCIP_NODE))
-        cdef SCIP_NODE* eqchild = <SCIP_NODE*> malloc(sizeof(SCIP_NODE))
-        cdef SCIP_NODE* upchild = <SCIP_NODE*> malloc(sizeof(SCIP_NODE))
+        cdef SCIP_NODE* downchild
+        cdef SCIP_NODE* eqchild
+        cdef SCIP_NODE* upchild
 
         PY_SCIP_CALL(SCIPbranchVarVal(self._scip, (<Variable>variable).scip_var, value, &downchild, &eqchild, &upchild))
-        # TODO should the stuff be freed and how?
+
         return Node.create(downchild), Node.create(eqchild), Node.create(upchild)
 
     def calcNodeselPriority(self, Variable variable, branchdir, targetvalue):
@@ -3390,7 +3400,7 @@ cdef class Model:
         :return: Node, the child which was created
 
         """
-        cdef SCIP_NODE* child = <SCIP_NODE*> malloc(sizeof(SCIP_NODE))
+        cdef SCIP_NODE* child
         PY_SCIP_CALL(SCIPcreateChild(self._scip, &child, nodeselprio, estimate))
         return Node.create(child)