+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 - add Model.getParams that returns a dict mapping all parameter names to their values
 - add Model.setParams to set multiple parameters at once using a dict
 - Add Model.from_ptr and Model.to_ptr to interface with SCIP* managed outside of PySCIPOpt
+- Node.getParent() returns None if the node has no parent
+- add Event.getRow() and Row.name
 
 ## 2.2.3 - 2019-12-10
 ### Added